fix apis

+++ app/Http/Controllers/api/v1/AuthenticationController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\api\v1;
 
 use App\Http\Controllers\Controller;
+use App\Models\Customer;
 use App\Models\Log as ModelsLog;
 use App\Models\User;
 use Illuminate\Http\Request;
@@ -16,64 +17,64 @@ use Stevebauman\Location\Facades\Location;
 
 class AuthenticationController extends Controller
 {
-    const ALLOWED_USER_TYPES = ["vendor","warehouse", "admin"];
-    
+    const ALLOWED_USER_TYPES = ["vendor", "warehouse", "admin", "customer"];
+
     public function signUp(Request $request)
     {
         try {
             $validator = Validator::make($request->all(), [
-                "firstName" => "required",
-                "lastName" => "required",
-                "phoneNumber" => "required|numeric|unique:tblvendor,phone",
-                "email" => "required|email|unique:tbluser,email",
+                "fname" => "required",
+                "lname" => "required",
+                "phone" => "required|numeric|unique:tblcustomer,phone",
+                // "email" => "required|email|unique:tbluser,email",
                 "password" => "required|min:8",
-                "username" => "unique:tblvendor,username",
-                "idType" => "required",
-                "idNumber" => "required",
+                // "username" => "unique:tblvendor,username",
+                "id_type" => "required",
+                "id_no" => "required",
                 // "idFileLink" => "required",
-                "region" => "required",
-                "town" => "required",
-                "streetName" => "required",
-                "landmark" => "required",
+                // "region" => "required",
+                // "town" => "required",
+                // "streetName" => "required",
+                // "landmark" => "required",
                 "gpsaddress" => "required",
             ], [
                 // This has our own custom error messages for each validation
-                "firstName.required" => "No first name supplied",
-                "lastName.required" => "No last name supplied",
+                "fname.required" => "No first name supplied",
+                "lname.required" => "No last name supplied",
 
                 // Usertype errors
-                "userType.required" => "No user type specified. We expect camel case 'userType' with a valid value",
-                "userType.in" => "Invalid usertype: expected one of [" . join(",", self::ALLOWED_USER_TYPES) . "] but got [{$request->userType}]",
+                // "userType.required" => "No user type specified. We expect camel case 'userType' with a valid value",
+                // "userType.in" => "Invalid usertype: expected one of [" . join(",", self::ALLOWED_USER_TYPES) . "] but got [{$request->userType}]",
 
                 // Phone error messages
-                "phoneNumber.required" => "No phone number supplied",
-                "phoneNumber.numeric" => "Phone number supplied [{$request->phoneNumber}] must contain only numbers",
-                "phoneNumber.unique" => "Phone number already taken",
+                "phone.required" => "No phone number supplied",
+                "phone.numeric" => "Phone number supplied [{$request->phoneNumber}] must contain only numbers",
+                "phone.unique" => "Phone number already taken",
 
                 // Gender error messages
-                "gender.required" => "Your gender was not specified",
-                "gender.max" => "Invalid value for gender: expected one of ['m','f'] but got [{$request->gender}]",
+                // "gender.required" => "Your gender was not specified",
+                // "gender.max" => "Invalid value for gender: expected one of ['m','f'] but got [{$request->gender}]",
 
                 // Email error messages
-                "email.email" => "The supplied email [{$request->email}] is not a valid email",
-                "email.required" => "No email supplied",
-                "email.unique" => "Email already taken",
+                // "email.email" => "The supplied email [{$request->email}] is not a valid email",
+                // "email.required" => "No email supplied",
+                // "email.unique" => "Email already taken",
 
-                 // Username error messages
-                 "username.unique" => "Username already taken",
+                // Username error messages
+                //  "username.unique" => "Username already taken",
 
-                 // ID error messages
-                 "idType.required" => "ID Type is required",
-                 "idNumber.required" => "ID number is required",
-                //  "idFileLink.required" => "Upload ID",
+                // ID error messages
+                "id_type.required" => "ID Type is required",
+                "id_no.required" => "ID number is required",
+                "id_link.required" => "Upload ID",
 
-                 // Address error messages
-                 "region.required" => "Region is required",
-                 "town.required" => "Town is required",
-                 "streetname.required" => "Streetname is required",
-                 "landmark.required" => "Landmark is required",
-                 "gpsaddress.required" => "GPS Address is required",
-                
+                // Address error messages
+                //  "region.required" => "Region is required",
+                //  "town.required" => "Town is required",
+                //  "streetname.required" => "Streetname is required",
+                //  "landmark.required" => "Landmark is required",
+                "gpsaddress.required" => "GPS Address is required",
+
                 // Password error messages
                 "password.required" => "No password supplied",
                 // "password.confirm" => "Your passwords do not match",
@@ -84,20 +85,20 @@ class AuthenticationController extends Controller
                 return response()->json([
                     "ok" => false,
                     "msg" => "Registration failed. " . join(". ", $validator->errors()->all()),
-                ]);
+                ], 422);
             }
 
             DB::beginTransaction();
             $transid = strtoupper(bin2hex(random_bytes(4)));
             DB::table("tbluser")->insert([
                 "transid" => $transid,
-                "userid" => 'VND-' . $transid,
-                "fname" => $request->firstName,
-                "lname" => $request->lastName,
-                "username" => $request->username,
-                "usertype" => "vendor",
+                "userid" => 'CUST-' . $transid,
+                "fname" => $request->fname,
+                "lname" => $request->lname,
+                "username" => $request->fname . ' ' . $request->lname,
+                "usertype" => "customer",
                 "password" =>  Hash::make($request->password),
-                "phone" => empty($request->phoneNumber) ? '' : $request->phoneNumber,
+                "phone" => empty($request->phone) ? '' : $request->phone,
                 "email" => $request->email,
                 "picture" => $request->picture,
                 "deleted" =>  0,
@@ -105,46 +106,56 @@ class AuthenticationController extends Controller
                 "createuser" =>  $request->createuser,
             ]);
 
-            DB::table('tblvendor')->insert([
-                "transid" => $transid,
-                "vendor_no" => 'VND-' . $transid,
-                "fname" => strtoupper($request->firstName),
-                "lname" => strtoupper($request->lastName),
-                "mname" => strtoupper($request->middleName),
-                "phone" => $request->phoneNumber,
-                "username" => $request->username,
-                "email" => $request->email,
-                "region" => $request->region,
-                "town" => $request->town,
-                "streetname" => $request->streetName,
-                "landmark" => $request->landmark,
-                "gpsaddress" => $request->gpsaddress,
-                "longitude" => $request->longitude,
-                "latitude" => $request->latitude,
-                "gender" => strtoupper($request->gender),
-                "dob" => $request->dateOfBirth,
-                "picture" => $request->picture,
-                "id_type" => $request->idType,
-                "id_no" => $request->idNumber,
-                "id_file_link" => $request->idFileLink,
-                "createdate" =>  date('Y-m-d H:i:s'),
-                "createuser" => $request->createuser,
-                "deleted" => 0,
-                "approved" => 0,
-            ]);
+            // DB::table('tblvendor')->insert([
+            //     "transid" => $transid,
+            //     "vendor_no" => 'VND-' . $transid,
+            //     "fname" => strtoupper($request->firstName),
+            //     "lname" => strtoupper($request->lastName),
+            //     "mname" => strtoupper($request->middleName),
+            //     "phone" => $request->phoneNumber,
+            //     "username" => $request->username,
+            //     "email" => $request->email,
+            //     "region" => $request->region,
+            //     "town" => $request->town,
+            //     "streetname" => $request->streetName,
+            //     "landmark" => $request->landmark,
+            //     "gpsaddress" => $request->gpsaddress,
+            //     "longitude" => $request->longitude,
+            //     "latitude" => $request->latitude,
+            //     "gender" => strtoupper($request->gender),
+            //     "dob" => $request->dateOfBirth,
+            //     "picture" => $request->picture,
+            //     "id_type" => $request->idType,
+            //     "id_no" => $request->idNumber,
+            //     "id_file_link" => $request->idFileLink,
+            //     "createdate" =>  date('Y-m-d H:i:s'),
+            //     "createuser" => $request->createuser,
+            //     "deleted" => 0,
+            //     "approved" => 0,
+            // ]);
 
+            $custno = 'CUST-' . $transid;
+            Customer::create($validator->validated() + ['custno' => $custno]);
+
+            if (null !== $request->file("id_link")) {
+                $filePath = $request->file("id_link")->store("public/student");
+                Customer::where("custno", $custno)->update([
+                    "id_link" => env("APP_URL") . "/" . str_replace("public", "storage", $filePath),
+                ]);
+            }
+            
             $userIp = $request->ip();
             $locationData = Location::get($userIp);
             $transid1 = strtoupper(bin2hex(random_bytes(4)));
 
             ModelsLog::insert([
                 "transid" => $transid1,
-                "username" => $request->createuser,
-                "module" => "Vendor",
+                "username" => $request->fname . ' ' . $request->lname,
+                "module" => "Customer",
                 "action" => "Add",
-                "activity" => "Vendor registered from Mobile with id VND-{$transid} successfully",
+                "activity" => "Customer registered from Mobile with id CUST-{$transid} successfully",
                 "ipaddress" => $userIp,
-                "createuser" =>  $request->createuser,
+                "createuser" =>  $request->fname . ' ' . $request->lname,
                 "createdate" => gmdate("Y-m-d H:i:s"),
                 "longitude" => $locationData->longitude ?? $userIp,
                 "latitude" => $locationData->latitude ?? $userIp,
@@ -155,7 +166,7 @@ class AuthenticationController extends Controller
             return response()->json([
                 "ok" => true,
                 "msg" => "Registration successful",
-            ]);
+            ], 200);
         } catch (\Throwable $e) {
             DB::rollBack();
             Log::error("An error occured during signup", [
@@ -175,12 +186,14 @@ class AuthenticationController extends Controller
     {
         try {
             $validator = Validator::make($request->all(), [
-                "email" => "required|email|exists:tbluser,email",
+                // "email" => "required|email|exists:tbluser,email",
                 "password" => "required",
+                "phone" => "required",
             ], [
-                "email.required" => "Email not supplied",
-                "email.email" => "Invalid email supplied: [{$request->email}]",
-                "email.exists" => "Invalid credentials",
+                // "email.required" => "Email not supplied",
+                // "email.email" => "Invalid email supplied: [{$request->email}]",
+                // "email.exists" => "Invalid credentials",
+                "phone.required" => "Phone number not supplied",
                 "password.required" => "Password not supplied",
             ]);
 
@@ -188,24 +201,24 @@ class AuthenticationController extends Controller
                 return response()->json([
                     "ok" => false,
                     "msg" => "Login failed. " . join(". ", $validator->errors()->all()),
-                ]);
+                ], 422);
             }
 
-            if (!Auth::attempt($request->only(["email", "password"]))) {
+            if (!Auth::attempt($request->only(["phone", "password"]))) {
                 return response()->json([
                     "ok" => false,
                     "msg" => "Login failed. Invalid credentials",
-                ]);
+                ], 418);
             }
 
-            $user = User::where("email", $request->email)->first();
+            $user = User::where("phone", $request->phone)->first();
 
             // No admin can login via the mobile
             if (!in_array(strtolower($user->usertype), self::ALLOWED_USER_TYPES)) {
                 return response()->json([
                     "ok" => false,
                     "msg" => "You cannot log in using the mobile client",
-                ]);
+                ], 418);
             }
 
             // Make sure account has been approved
@@ -217,7 +230,10 @@ class AuthenticationController extends Controller
             // }
 
             $token = $user->createToken('accessToken')->plainTextToken;
-            
+            User::where("phone", $request->phone)->update([
+                'remember_token' => $token
+            ]);
+
             $userIp = $request->ip();
             $locationData = Location::get($userIp);
             $transid1 = strtoupper(bin2hex(random_bytes(4)));
@@ -249,35 +265,31 @@ class AuthenticationController extends Controller
                 "msg" => $e->getMessage(),
                 "trace" => $e->getTrace(),
             ]);
-
             return response()->json([
                 "ok" => false,
                 "msg" => "An internal error occured",
                 "trace" => $e->getTrace(),
-            ]);
+            ], 500);
         }
     }
-    
+
     public function logout(Request $request)
     {
-        try {
-        } catch (\Throwable $e) {
-            Log::error("Error logging out", [
-                "errMsg" => $e->getMessage(),
-                "trace" => $e->getTrace(),
-            ]);
+        if ($request->user()) {
+            $request->user()->currentAccessToken()->delete();
             return response()->json([
-                "ok" => false,
-                "msg" => "An internal error"
+                "ok" => true,
+                "msg" => "Logout successful.",
             ]);
         }
-        $request->user()->currentAccessToken()->delete();
-        
+
+        // If the user is not authenticated, return an unauthorized error response
         return response()->json([
-            "ok" => true,
-            "msg" => "Logged out successfully",
-        ]);
+            "ok" => false,
+            "msg" => "No authenticated user found.",
+        ], 401);
     }
+
 
     /**
      * Allows user to change their password
@@ -366,7 +378,7 @@ class AuthenticationController extends Controller
                 "longitude" => $locationData->longitude ?? $userIp,
                 "latitude" => $locationData->latitude ?? $userIp,
             ]);
-            
+
             return response()->json([
                 "ok" => true,
                 "msg" => "Password successfully changed",
@@ -397,11 +409,11 @@ class AuthenticationController extends Controller
     public function logReport($userid, $dateFrom, $dateTo)
     {
         $logs = DB::table("tbllogs")
-        ->when($userid !== 'all', function ($q)  use ($userid) {
-            return $q->where('userid', $userid);
-        })
-        ->whereBetween('createdate', [$dateFrom, $dateTo])
-        ->orderByDesc("createdate")->get();
+            ->when($userid !== 'all', function ($q)  use ($userid) {
+                return $q->where('userid', $userid);
+            })
+            ->whereBetween('createdate', [$dateFrom, $dateTo])
+            ->orderByDesc("createdate")->get();
 
         return response()->json([
             "data" => $logs
